+++ sampledb/frontend/admin_warnings.py
@@ -5,7 +5,7 @@ import flask
 import flask_login
 
 from . import frontend
-from ..logic.utils import show_load_objects_in_background_warning, show_admin_local_storage_warning
+from ..logic.utils import show_load_objects_in_background_warning, show_admin_local_storage_warning, show_numeric_tags_warning, do_numeric_tags_exist
 
 
 @frontend.route('/admin/warnings/')
@@ -17,6 +17,7 @@ def admin_warnings():
     for deprecation_warning_check in [
         show_load_objects_in_background_warning,
         show_admin_local_storage_warning,
+        show_numeric_tags_warning,
     ]:
         warnings[deprecation_warning_check.__name__] = deprecation_warning_check()
 
@@ -24,5 +25,6 @@ def admin_warnings():
     return flask.render_template(
         'admin/warnings.html',
         has_warnings=has_warnings,
+        do_numeric_tags_exist=do_numeric_tags_exist() if show_numeric_tags_warning() else False,
         **warnings
     )